Refactored Book and Fillable out of hist.py
