+++ histbook/version.py
@@ -30,7 +30,7 @@
 
 import re
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 version = __version__
 version_info = tuple(re.split(r"[-\.]", __version__))
